Improved ObjectPool class
Fixed a bug where I was creating extra GameObjects in the scene in
ParticlePool::Awake(), because of the way I was creating a transform.

To fix this I added more Create() methods to the ObjectPool base class
to negate the need for overloads in the derived classes; CubePool and
ParticlePool. This made it easy to set object positions when spawning
objects.